improve code for com_rockcarry_ffplayer_player.cpp.

+++ player-android/jni/com_rockcarry_ffplayer_player.cpp
@@ -81,7 +81,7 @@ JNIEXPORT jlong JNICALL Java_com_rockcarry_ffplayer_player_nativeOpen
   (JNIEnv *env, jclass clazz, jstring url, jobject jsurface, jint w, jint h)
 {
     const char *str = env->GetStringUTFChars(url, NULL);
-    jint hplayer = (jint)player_open((char*)str, NULL);
+    jlong hplayer = (jlong)player_open((char*)str, NULL);
     env->ReleaseStringUTFChars(url, str);
     return hplayer;
 }
